# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Using AI to do something.",
        "title": "AI Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/wallet/create": {
            "post": {
                "description": "Create wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create wallet",
                "operationId": "CreateWallet",
                "parameters": [
                    {
                        "description": "Wallet Create Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.CreatedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wallet.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/wallet/delete/{id}": {
            "delete": {
                "description": "Delete wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Delete wallet",
                "operationId": "DeleteWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wallet.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/wallet/detail/{id}": {
            "get": {
                "description": "Detail wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Detail wallet",
                "operationId": "DetailWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wallet.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/wallet/list": {
            "get": {
                "description": "List wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "List wallet",
                "operationId": "ListWallet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "FilterName",
                        "name": "filterName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortFields",
                        "name": "sortFields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Dir",
                        "name": "dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wallet.ResponseList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/wallet/update/{id}": {
            "put": {
                "description": "Update wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Update wallet",
                "operationId": "UpdateWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Wallet Update Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.UpdatedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wallet.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Chain": {
            "type": "string",
            "enum": [
                "None",
                "Bitcoin",
                "Ethereum",
                "Polygon"
            ],
            "x-enum-varnames": [
                "None",
                "Bitcoin",
                "Ethereum",
                "Polygon"
            ]
        },
...
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "wallet.BalancesResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "currencyId": {
                    "type": "string"
                },
                "decimal": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "wallet.CreatedRequest": {
            "type": "object",
            "required": [
                "chain",
                "description",
                "name",
                "userId"
            ],
            "properties": {
                "chain": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Chain"
                        }
                    ],
                    "example": "Polygon"
                },
                "description": {
                    "type": "string",
                    "example": "this is for admin wallet"
                },
                "name": {
                    "type": "string",
                    "example": "admin"
                },
                "userId": {
                    "type": "string",
                    "example": "abcdef2nopabcdef2nop"
                }
            }
        },
        "wallet.Response": {
            "type": "object",
            "required": [
                "chain",
                "userId"
            ],
            "properties": {
                "chain": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Chain"
                        }
                    ],
                    "example": "Polygon"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string",
                    "example": "abcd-efgh-ijkl-mnop"
                },
                "walletBalances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wallet.BalancesResponse"
                    }
                }
            }
        },
        "wallet.ResponseList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wallet.Response"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "wallet.UpdatedRequest": {
            "type": "object",
            "required": [
                "chain",
                "description",
                "name",
                "userId"
            ],
            "properties": {
                "chain": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Chain"
                        }
                    ],
                    "example": "Polygon"
                },
                "description": {
                    "type": "string",
                    "example": "this is for admin wallet"
                },
                "name": {
                    "type": "string",
                    "example": "admin"
                },
                "userId": {
                    "type": "string",
                    "example": "abcd-efgh-ijkl-mnop"
                }
            }
        }
    }
}